+++ sockeye/arguments.py
@@ -767,12 +767,12 @@ def add_training_args(params):
                               help="The MXNet kvstore to use. 'device' is recommended for single process training. "
                                    "Use any of 'dist_sync', 'dist_device_sync' and 'dist_async' for distributed "
                                    "training. Default: %(default)s.")
-    train_params.add_argument("--gctype",
+    train_params.add_argument("--gradient-compression-type",
                               type=str,
                               default=None,
                               choices=[None, "2bit"],
                               help='Type of gradient compression to use. Choose "2bit" or None. Default: %(default)s.')
-    train_params.add_argument("--gcthreshold",
+    train_params.add_argument("--gradient-compression-threshold",
                               type=float,
                               default=0.5,
                               help="Threshold for gradient compression if --gctype is '2bit'. Default: %(default)s.")